+++ src/hooks.ts
@@ -60,6 +60,7 @@ export const HOOKS: { [k: string]: any[]; } = {
      * Notice that you have to clone stuff before mutating the inputs.
      */
     preCreateRxStorageInstance: [],
+    preStorageWrite: [],
     /**
      * runs on the document-data before the document is migrated
      * {